+++ test/fixture/example/invalid-values.sh
@@ -4,7 +4,7 @@ export EARTH_ATOM_COUNT=5.9722e24
 export GRPC_TIMEOUT=10S
 export LOG_LEVEL=silly
 export PORT=65536
-export READ_DSN=
+export READ_DSN="host=localhost"
 export REDIS_PRIMARY_SERVICE_HOST=.redis.example.org
 export REDIS_PRIMARY_SERVICE_PORT=65536
 export SAMPLE_RATIO=1/100